Write and pass test for 3rd vertical win condition
The code duplication has been noted, but refactoring is going to be delayed until post diagonal win conditions to make the pattern more apparent.

I need to be more diligent in test reading. A red bar doesn't mean "go", it means "read why it's red", I have been making too many typos in my test and just assuming that I was good to go because hey, I expected it not to pass and it didn't pass.