Pass ai can move test, add test for moving to empty spot
While I recognize that this test and the previous one can now be passed by changing the implementation of get_move in ai to just return 8, writing more tests doesn't seem like the right move. It neither clarifies nor forces me to do something I wouldn't already be doing to pass this test if I weren't actively trying to be a pedant.