Add test for winning move, change implementation of empty spot
So, I decided to go the hard-code route and change the return value of "get_move" from a random value that hasn't already been taken to a hardcoded 0 (changed from 8 for readability reasons).

I did this because I realized that the implementation I had with the random return _could_ pass the new test I wrote and I'd rather have a failing test than a flapping test, even if I know why the test is flapping.