Refactor win conditions to exist in separate module
I don't love the idea of the board knowing/calculating its own win conditions and both Game and AI need access to it, putting it into a separate module allows Game and AI to access it without giving it directly to board. That being said, the import line is duplicated.