Change board marking from marker to turn
Instead of reflecting the marker of the player, the board now marks which turn a spot was taken. Eventually this will allow for the players to stop holding onto their past turns. Furthermore, markers, colors and board color is now held in an object of the Aesthetics class. This class still needs to be properly implemented.