Modify create_board route to use newer game functionality
Use the game instead of the board, this will allow the game to handle turns later, cutting down on turn and player logic happening in the routes.

Add one more assert to a board test.

Old code has been left in as comments to guide the refactor and clean-up will happen later.