Pull out helper methods for routes, error handle, redirect
Instead of Flask's default panic ("bad request") message when an incomplete form is submit, it now redirects to the form with an error message. One shouldn't be able to jump into a board or game over message now without having played some of the game (specified routes are so Flask doesn't try redirecting the css).

Helper methods are now in their own file as a step toward SRP.

game over and vs. machine doesn't currently work - in progress.