Replace call to a method that doesn't exist, fix test
Method call for self._copy_board_for_speed was an artifact from a spike, it has been replaced with the native copy function.

Due to a fundamental misunderstanding on my part about how to play tictactoe with any skill, "test_move_logic_prevents_fork_16_space_board" was missing a condition that would have prevented the fork - or more accurately, if the other player had created a fork, it would have led to the AI winning anyway. This additional option has been added as a passing condition.